requirements.txt
langchain>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.22.0
tiktoken>=0.4.0
faiss-cpu>=1.7.0
protobuf>=3.19.0
langchain-experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb
requests>=2.28.0
# Optional: Hugging Face dependencies (install separately if needed)
# transformers>=4.21.0
# torch>=1.9.0
# accelerate>=0.20.0
# langchain-huggingface